Disable QuickTime in OpenCV for MacOS Sierra and later.
The QuickTime framework is no longer provided as of MacOS 10.12 so this
change disables QTKit in OpenCV when building on newer versions of
MacOS.  Upstream OpenCV has fixed this issue by adding support for
AVFoundation, which is the QuickTime API replacement.  This change just
disables the broken QTKit option until we upgrade OpenCV in Fletch.