Switch motif to building minimal set of Qt packages.
Previously we were selecting webkit on/off and javascript on/off. With the build failures with gcc >= 6 we were faced with another removal, script and scripttools. It makes more sense to turn off everything or all all to build. Turning those packages on was a use case that was almost never activated.