Fix PROJ on Windows
For some reason, on Windows, PROJ inconsistently installs some things to
local/{lib,include} instead of just {lib,include} as on every other
platform. This breaks users due to the way we "front load"
PROJ_INCLUDE_DIR. While we could adjust that to have the correct
location depending on platform, it seems easier to just force the
"usual" locations for all platforms. (On not-Windows, this is
superfluous, but saves fiddling with platform-dependent CMake
arguments.)