Patch to allow OpenCV to build with Python on MSVC in Debug
There is an issue with pyconfig.h provided by Python in that if you are
building on Windows it will force your executable to link to either
python27.lib or python27_d.lib depending on whether it is built in Debug
mode or not.  At the same time, python27_d.lib is not distributed with
Python and only available if you build Python from source, which is quite
rare on Windows.  This patch tricks Python into linking to python27.lib
for Debug builds by temporarily unsetting _DEBUG.  When including the
Python.h header.

A slightly better fix would use CMake to determine if the Python debug
libraries are found and then only enable this hack if they are not found.