Set the PROJ data directory consistently
For whatever reason PROJ polutes the root of the "share" directory
with its data files on Windows only, but puts them in "share/proj"
on other platforms.  We need them in a consistent subdirectory to be
able to later figure out which files came from PROJ when building
installer packages.  This changes fixes the data directory to
"share/proj" on all platforms.