Improve GeographicLib targets
Patch GeographicLib to generate a utility (INTERFACE) target that links
to either the shared or static library, depending on availability. This
avoids consumers needing to care about which one was built in order to
use targets for linking.

Upstream seems potentially receptive to this, although we probably will
not be able to upgrade our version of GeographicLib immediately. Also,
adding this to Fletch first gives us a way to verify that the approach
works, which should be beneficial to upstream acceptance.