"Always" allow Qt version selection
Modify how we set up the Qt version selection to show it if any
Qt-dependent package is being built, whether or not we are actually
building Qt. This allows users to select which version of Qt they want
for Qt-dependent packages while still using an external Qt.

For now, this means VTK, but the next commit will expand this to
qtExtensions also.