Fix the setting the Python executable path for Qt.
Qt requires Python on the system path to build QtQml or the build with fail
with "Building QtQml requires Python".  This code was previously inserting
the full path to the Python executable into the PATH varible.
Instead we need to add the directory where the Python executable is found.

This changes fixes build errors on Windows when Python is not globally in
the system PATH.