Set library path when building Qt
Because Qt builds tools that are used to build other parts of Qt, some
of which may need libraries that Qt was built against, we may need to
fiddle with the environment so that these libraries can be found. Add
logic to do that for Windows. Also, tweak the GDAL logic that was doing
something similar to use `cmake -E env` rather than `env` to do its
environment manipulation (this is more portable and easier to specify).