Pass BOOST_INCLUDE_DIR instead of BOOST_ROOT
The BOOST_ROOT variable has problems with CMake master because of a
change in the FindBoost module. The module will fatal error now if
BOOST_ROOT is set and cmake_policy CMP0074 is OLD. This issue is a
possible cmake bug but this fix is sound since BOOST_INCLUDE_DIR is
sufficient for finding Boost.