Move the nasm args into the already patched CMakeList.txt for simd
The clear intent of the original CMakeLists.txt that we patched was to
run nasm(yasm) with -g in Debug or RelWithDebInfo. We removed those
lines and added -g v8 to the CMake arguments. The new CMake 3.20.1 is
not happy with those args added in the way we were doing it so
libjpeg-turbo stopped configuring. This patch moves those arguments into
its expected location and condition."