Remove the advanced marking for the fletch_ENABLE_VTK_PYTHON option.
Keeping this option enabled can cause build issues on some platforms and adds a lot of time to the build, especially in Visual Studio. I don't see why it should be marked as advanced. It is better to have the option exposed so folks can remember that it's there before taking so much time building it. VTK is about the last build to finish because of Qt so seeing the failure at the end can be painful.